refactor(soup): rename CallChannelName to CallRecordName
The entity already carries a resolved name (custom_name or channel
name), so the component no longer needs the channel-preview fallback.
Drop CallChannelNameFromPreview, simplify rendering, rename to
CallRecordName, and align the search transform's default fallback
with soup (blockNameToDefaultFile('call')).
